+++ test/over_react/component_declaration/builder_integration_tests/component2/annotation_error_integration_test.over_react.g.dart
@@ -51,16 +51,12 @@ class _$$AnnotationErrorDefaultPropsProps
     extends _$AnnotationErrorDefaultPropsProps
     with _$AnnotationErrorDefaultPropsPropsAccessorsMixin
     implements AnnotationErrorDefaultPropsProps {
-  // This initializer of `_props` to an empty map, as well as the reassignment
-  // of `_props` in the constructor body is necessary to work around a DDC bug: https://github.com/dart-lang/sdk/issues/36217
-  _$$AnnotationErrorDefaultPropsProps(Map? backingMap) : this._props = {} {
-    this._props = backingMap ?? {};
-  }
+  _$$AnnotationErrorDefaultPropsProps([Map? backingMap])
+      : this.props = backingMap ?? JsBackedMap();
 
   /// The backing props map proxied by this class.
   @override
-  Map get props => _props;
-  Map _props;
+  final Map props;
 
   /// Let `UiProps` internals know that this class has been generated.
   @override
@@ -153,16 +149,12 @@ _$$AnnotationErrorProps _$AnnotationError([Map? backingProps]) =>
 class _$$AnnotationErrorProps extends _$AnnotationErrorProps
     with _$AnnotationErrorPropsAccessorsMixin
     implements AnnotationErrorProps {
-  // This initializer of `_props` to an empty map, as well as the reassignment
-  // of `_props` in the constructor body is necessary to work around a DDC bug: https://github.com/dart-lang/sdk/issues/36217
-  _$$AnnotationErrorProps(Map? backingMap) : this._props = {} {
-    this._props = backingMap ?? {};
-  }
+  _$$AnnotationErrorProps([Map? backingMap])
+      : this.props = backingMap ?? JsBackedMap();
 
   /// The backing props map proxied by this class.
   @override
-  Map get props => _props;
-  Map _props;
+  final Map props;
 
   /// Let `UiProps` internals know that this class has been generated.
   @override
@@ -253,16 +245,12 @@ _$$AnnotationErrorStatefulProps _$AnnotationErrorStateful(
 class _$$AnnotationErrorStatefulProps extends _$AnnotationErrorStatefulProps
     with _$AnnotationErrorStatefulPropsAccessorsMixin
     implements AnnotationErrorStatefulProps {
-  // This initializer of `_props` to an empty map, as well as the reassignment
-  // of `_props` in the constructor body is necessary to work around a DDC bug: https://github.com/dart-lang/sdk/issues/36217
-  _$$AnnotationErrorStatefulProps(Map? backingMap) : this._props = {} {
-    this._props = backingMap ?? {};
-  }
+  _$$AnnotationErrorStatefulProps([Map? backingMap])
+      : this.props = backingMap ?? JsBackedMap();
 
   /// The backing props map proxied by this class.
   @override
-  Map get props => _props;
-  Map _props;
+  final Map props;
 
   /// Let `UiProps` internals know that this class has been generated.
   @override
@@ -313,16 +301,12 @@ class AnnotationErrorStatefulState extends _$AnnotationErrorStatefulState
 class _$$AnnotationErrorStatefulState extends _$AnnotationErrorStatefulState
     with _$AnnotationErrorStatefulStateAccessorsMixin
     implements AnnotationErrorStatefulState {
-  // This initializer of `_state` to an empty map, as well as the reassignment
-  // of `_state` in the constructor body is necessary to work around a DDC bug: https://github.com/dart-lang/sdk/issues/36217
-  _$$AnnotationErrorStatefulState(Map? backingMap) : this._state = {} {
-    this._state = backingMap ?? {};
-  }
+  _$$AnnotationErrorStatefulState([Map? backingMap])
+      : this.state = backingMap ?? JsBackedMap();
 
   /// The backing state map proxied by this class.
   @override
-  Map get state => _state;
-  Map _state;
+  final Map state;
 
   /// Let `UiState` internals know that this class has been generated.
   @override
@@ -403,17 +387,12 @@ class _$$AnnotationErrorStatefulDefaultPropsProps
     extends _$AnnotationErrorStatefulDefaultPropsProps
     with _$AnnotationErrorStatefulDefaultPropsPropsAccessorsMixin
     implements AnnotationErrorStatefulDefaultPropsProps {
-  // This initializer of `_props` to an empty map, as well as the reassignment
-  // of `_props` in the constructor body is necessary to work around a DDC bug: https://github.com/dart-lang/sdk/issues/36217
-  _$$AnnotationErrorStatefulDefaultPropsProps(Map? backingMap)
-      : this._props = {} {
-    this._props = backingMap ?? {};
-  }
+  _$$AnnotationErrorStatefulDefaultPropsProps([Map? backingMap])
+      : this.props = backingMap ?? JsBackedMap();
 
   /// The backing props map proxied by this class.
   @override
-  Map get props => _props;
-  Map _props;
+  final Map props;
 
   /// Let `UiProps` internals know that this class has been generated.
   @override
@@ -471,17 +450,12 @@ class _$$AnnotationErrorStatefulDefaultPropsState
     extends _$AnnotationErrorStatefulDefaultPropsState
     with _$AnnotationErrorStatefulDefaultPropsStateAccessorsMixin
     implements AnnotationErrorStatefulDefaultPropsState {
-  // This initializer of `_state` to an empty map, as well as the reassignment
-  // of `_state` in the constructor body is necessary to work around a DDC bug: https://github.com/dart-lang/sdk/issues/36217
-  _$$AnnotationErrorStatefulDefaultPropsState(Map? backingMap)
-      : this._state = {} {
-    this._state = backingMap ?? {};
-  }
+  _$$AnnotationErrorStatefulDefaultPropsState([Map? backingMap])
+      : this.state = backingMap ?? JsBackedMap();
 
   /// The backing state map proxied by this class.
   @override
-  Map get state => _state;
-  Map _state;
+  final Map state;
 
   /// Let `UiState` internals know that this class has been generated.
   @override