+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # OverReact Changelog
 
 ## Unreleased
+- [#989] Optimize generated code to decrease dart2js compile size, saving ~577 bytes per component (when using `-03 --csp --minify`)
 - [#992] Fix compilation errors for legacy boilerplate defined in libraries with a Dart language version of >=3.0
 
 ## 5.4.6